+++ CopilotKit/packages/runtime-client-gql/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @copilotkit/runtime-client-gql
 
+## 1.8.6
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.6
+
+## 1.8.6-next.0
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.6-next.0
+
+## 1.8.5
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.5
+
+## 1.8.5-next.5
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.5-next.5
+
+## 1.8.5-next.4
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.5-next.4
+
 ## 1.8.5-next.3
 
 ### Patch Changes